Add files via upload
This is a model that predicts which students are likely to enroll in a particular program and which ones may need additional support to graduate .I have provided the csv file that i used for training and also checking  the functionality of the model. Then i  have added some screenshots showing the output in the terminal also included is an output csv predicting the predicted results which is used to communicate results of the model to the educational institutions.